Refactor: Lotto.jsfmf  src/로 이동
