+++ src/Models/index.js
@@ -1,3 +1,3 @@
 export { default as DrawnLotto } from "./DrawnLotto.js";
-export { default as Lotto } from "./Lotto.js";
+export { default as Lotto } from "../Lotto.js";
 export { default as PurchaseAmount } from "./PurchaseAmount.js";
